+++ NEWS.md
@@ -4,6 +4,8 @@ ClimaCore.jl Release Notes
 main
 -------
 
+- Add support for bilinear interpolation for diagnostics to the Remapper. [2455](https://github.com/CliMA/ClimaCore.jl/pull/2455)
+
 v0.14.49
 -------
 - Add `PressureInterpolator` [2422](https://github.com/CliMA/ClimaCore.jl/pull/2422)